feat: port zlib, libpng and freetype
- switch to LLVM/Clang from GCC to allow for easier cross-compilation
- split libstd into libm and libc to make porting easier
- implement needed libm functions for new ports

+++ comp/libraries/libdraw/src/vertices.c
@@ -18,8 +18,8 @@ void vertices_circle(float* vertices, size_t count, float centerX, float centerY
     {
         end = start + 2.0F * M_PI;
     }
- 
-    if (start == 0.0f && fabsf(end - 2.0F * M_PI) < 0.0001f)
+
+    if (start == 0.0f && fabsf(end - 2.0F * (float)M_PI) < 0.0001f)
     {
         float step = (end - start) / (float)count;
         for (size_t i = 0; i < count; i++)
